+++ src/runners/check-prettier/run.js
@@ -1,0 +1,10 @@
+//--------------------------------------------------------
+//-- Check Prettier runner - Run
+//--------------------------------------------------------
+import prettier from "../../helpers/prettier.js";
+
+const checkPrettierRunner = ({ testPath }) => {
+	return prettier.run(testPath);
+};
+
+export default checkPrettierRunner;