+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Validate Prettier formatting and configuration
 
 ### Changed
+- ESLint config update to 3.0.0-beta.3
 - Maintenance update
 
 ### Fixed